pyproject.toml
[project]
name = "iesb"
version = "0.0.1"
description = "iesb"
readme = "README.md"
license = "MIT"
authors = [{ name = "Izaac", email = "izaacjrrm@email.com" }]
requires-python = ">=3.13.0"
dependencies = [
    "aiosqlite>=0.21.0",
    "colorlog>=6.8.0",
    "discord.py>=2.3.0",
    "pydantic>=2.11.7", 
    "python-decouple>=3.8",
    "sqlalchemy>=2.0.43",
    "tinydb>=4.8.2",
]

[project.urls]

[project.optional-dependencies]
dev = ["ruff", "pyright", "pytest", "pytest-xdist"]

[project.scripts]
iesb = "main:main"

[tool.ruff]
line-length = 88
target-version = "py313"
fix = true
show-fixes = true
indent-width = 4
exclude = ["venv", ".venv", "env", ".env", "node_modules", "__pycache__"]

[tool.ruff.lint]
select = [
    "ASYNC",
    "A",
    "B",
    "BLE",
    "C4",
    "C90",
    "COM",
    "E",
    "EM",
    "ERA",
    "EXE",
    "F",
    "FBT",
    "FIX",
    "I",
    "ICN",
    "ISC",
    "Q",
    "RET",
    "RSE",
    "S",
    "SIM",
    "SLF",
    "T10",
    "T20",
    "TC",
    "TD",
    "TRY",
    "UP",
    "W",
    "YTT",
    "RUF",
    "N"
]
ignore = [
    "FBT003",   # boolean args (APIs do Discord.py)
    "C901",     # complex function (tratamento de erros)
    "COM812",   # missing-trailing-comma (conflita com formatter)
    "E501",     # line too long (gerado automaticamente pelo Black)
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["ANN201", "S101"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"

[tool.ruff.lint.isort]
# Ensina ao Ruff qual é o pacote principal do seu projeto.
# Deve ser o mesmo nome da pasta dentro de 'src/'.
known-first-party = ["comandos"]


# ============================
# Tipagem (Pyright)
# ============================
[tool.pyright]
typeCheckingMode = "strict"
pythonVersion = "3.14"
include = ["src", 'tests']
exclude = [
    "**/venv",
    "**/.venv",
    "**/env",
    "**/.env",
    "**/node_modules",
    "**/__pycache__",
    "**/.github",
    "**/.vscode",
    "**/docs"
]
venv = ".venv"
venvPath = "."
executionEnvironments = [{ root = "src" }]
reportUnknownMemberType = false
reportUnknownParameterType = false

# ============================
# Testes (Pytest)
# ============================
[tool.pytest.ini_options]
addopts = "-s --color=yes --tb=short"
# Adiciona 'src' ao path do Python durante os testes. Isso permite que
# os testes em `tests/` importem os módulos de `src/` sem erros.
pythonpath = ["src"]
testpaths = ["tests"]


# ============================
# Build e Setuptools
# ============================
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[dependency-groups]
dev = [
    "pytest>=8.4.1",
]

[tool.setuptools.packages.find]
# Informa ao setuptools para encontrar os pacotes dentro da pasta 'src'.
# Essencial para a estrutura de layout com 'src/'.
where = ["src"]

[tool.setuptools]
# Mapeia o diretório raiz de pacotes para 'src'.
package-dir = { "" = "src" }
